+++ packages/connect-core/src/index.ts
@@ -16,6 +16,6 @@ export { default as Organization } from './entities/Organization'
 export { default as Permission, PermissionData } from './entities/Permission'
 export { default as Repo, RepoData } from './entities/Repo'
 export { default as Role, RoleData } from './entities/Role'
-export { ConnectionContext } from './types'
+export { ConnectionContext, IpfsResolver } from './types'
 export * from './utils'
 export * from './errors'
