+++ packages/connect-core/src/utils/address.ts
@@ -1,5 +1,6 @@
 import { Address } from '@aragon/connect-types'
-import { utils as ethersUtils } from 'ethers'
+import { providers as ethersProviders, utils as ethersUtils } from 'ethers'
+import { ErrorInvalidLocation } from '../errors'
 
 export const ANY_ENTITY = '0xFFfFfFffFFfffFFfFFfFFFFFffFFFffffFfFFFfF'
 
@@ -17,4 +18,25 @@ export function includesAddress(arr: Address[], address: Address): boolean {
 
 export function isAddress(address: string): boolean {
   return ethersUtils.isAddress(address)
+}
+
+export async function resolveAddress(
+  ethersProvider: ethersProviders.Provider,
+  location: string
+): Promise<Address> {
+  const isLocationAddress = isAddress(location)
+
+  const address = isLocationAddress
+    ? location
+    : await ethersProvider.resolveName(location)
+
+  if (!isAddress(address)) {
+    throw new ErrorInvalidLocation(
+      isLocationAddress
+        ? `The address (${address}) is not valid.`
+        : `The ENS domain (${location}) doesn’t seem to resolve to an address.`
+    )
+  }
+
+  return address
 }
