+++ packages/connect-core/src/utils/index.ts
@@ -1,6 +1,7 @@
 export * from './address'
 export * from './app-connectors'
 export * from './descriptions'
+export * from './ipfs'
 export * from './misc'
 export * from './network'
 export * from './subscriptions'
